+++ packages/apollo-bundle/src/ApolloBundle.ts
@@ -1,5 +1,5 @@
 import { ApolloServer, ApolloServerOptions, BaseContext } from "@apollo/server";
-import { expressMiddleware } from "@apollo/server/express4";
+import { expressMiddleware } from "@as-integrations/express4";
 import { ApolloServerPluginDrainHttpServer } from "@apollo/server/plugin/drainHttpServer";
 import {
   handlers,