+++ packages/logger-bundle/src/__tests__/ecosystem.ts
@@ -1,8 +1,9 @@
 import { Kernel } from "@bluelibs/core";
 import { LoggerBundle } from "../LoggerBundle";
+import { ILoggerBundleConfig } from "../defs";
 
-export const createKernel = (): Kernel => {
+export const createKernel = (config: ILoggerBundleConfig = {}): Kernel => {
   return new Kernel({
-    bundles: [new LoggerBundle()],
+    bundles: [new LoggerBundle(config)],
   });
 };