+++ packages/logger-bundle/src/listeners/ConsoleListener.ts
@@ -1,17 +1,25 @@
-import { Event, Listener, Service } from "@bluelibs/core";
+import { Listener, Service } from "@bluelibs/core";
 import { LogEvent } from "../events";
 import chalk, { Chalk } from "chalk";
-import { ILog, LogLevel } from "../defs";
+import { LogLevel, LogLevelOrder } from "../defs";
 
 @Service()
 export class ConsoleListener extends Listener {
   lastLogDate?: Date;
+  minLogLevel: LogLevel = LogLevel.DEBUG;
 
   init() {
-    this.on(LogEvent, (e: Event<{ log: ILog }>) => {
+    this.on(LogEvent, (e: LogEvent) => {
       const log = e.data.log;
 
-      let color: Chalk;
+      if (
+        LogLevelOrder.indexOf(log.level) >
+        LogLevelOrder.indexOf(this.minLogLevel)
+      ) {
+        return;
+      }
+
+      let color: Chalk = chalk.gray;
       // what are some good colors?
       if (log.level == LogLevel.INFO) {
         color = chalk.blueBright;